Bugfixes.
It now passes the regression test suite to the same degree as the
commit I started with.  Also, if you mechanically add
"from __future__ import co_annotations" to the top of
Lib/test/*py--with some exceptions!--it still passes.

Exceptions: there are a bunch of files that are sensitive to
line number changes, and some which are sensitive to the standard
loadout of futures.

Also, there are three legitimate crashes, which I'll (hopefully)
fix next: test_inspect, test_types, test_typing.