Fix bug in assigning co_annotations on a fn.
The bug: if a function had __annotations__ set to a dict,
and you set __co_annotations__ on a function object to None,
it would clear __annotations__.

This was lurking in the regression test suite, discovered
kind of by accident by Lib/tests/test_functools.py.
functools.wrap() (etc) copy attributes from the old function
to the new, and it was doing __annotations__ before
__co_annotations__.  If the order was reversed it wouldn't
have tripped the bug!