Memory optimization: tuple instead of function.
If the co_annotations function is a closure, or if it needs to
(possibly) refer to class names, we need an extra object when
we bind the function.  (Or, if both are true, we need two extra
objects.)  Rather than give up and bind the function immediately,
we now store all the objects (up to three) in a tuple, and store
that tuple as __co_annotations__.  It's only a small memory
savings, the tuple vs the full PyFunctionObject, but since most
annotations are never examined at runtime--and there can be a
lot of 'em!--this optimization seems worth doing.